Move "normalization" test data and dependent tests all into tests/
Leaves all other tests in place.

Provides a #[doc(hidden)] shim for accessing/using otherwise private
members from tests/
